move latin_hypercube
