--- xopt/generators/scipy/__init__.py
@@ -1,9 +1,0 @@
-from xopt.generators.scipy.latin_hypercube import LatinHypercubeGenerator
-from xopt.generators.sequential.neldermead import NelderMeadGenerator
-from xopt.generators.sequential.scipy import ScipyGenerator
-
-__all__ = [
-    "LatinHypercubeGenerator",
-    "NelderMeadGenerator",
-    "ScipyGenerator",
-]
